fix : 내가 작성한 Lite 서베이 진행중
