+++ src/app/(providers)/(root)/(private)/profile-post/[userId]/page.tsx
@@ -1,34 +1,48 @@
 'use client';
-import {getUserPostsIT} from '@/app/api/firebaseApi';
 import Link from 'next/link';
 import {useParams} from 'next/navigation';
 import {useEffect, useState} from 'react';
-interface Post {
+import {getUserPostLite, getUserPostsIT} from '../_components/getUserPost';
+interface PostIT {
   id: string;
   title: string;
   content: string;
   deadlineDate: Date | null;
 }
+interface PostLite {
+  id: string;
+  title: string;
+  content: string;
+}
 
 export default function ProfilePost() {
-  const [posts, setPosts] = useState<Post[]>([]);
+  const [posts, setPosts] = useState<PostIT[]>([]);
   const params = useParams<{userId: string}>();
   const userId = params.userId;
+  const [userPostLite, setUserPostLite] = useState<PostLite[]>([]);
 
   useEffect(() => {
     if (userId) {
       getUserPostsIT(userId).then(data => {
+        // 기존 게시글 데이터 설정
         setPosts(
-          data.map(doc => {
-            console.log(doc);
-            console.log(doc.id);
-            return {
-              id: doc.id,
-              title: doc.title,
-              content: doc.content,
-              deadlineDate: doc.deadlineDate,
-            };
-          }),
+          data.map(doc => ({
+            id: doc.id,
+            title: doc.title,
+            content: doc.content,
+            deadlineDate: doc.deadlineDate,
+          })),
+        );
+      });
+
+      getUserPostLite(userId).then(data => {
+        // 새로운 게시글 데이터 설정
+        setUserPostLite(
+          data.map(doc => ({
+            id: doc.id,
+            title: doc.title,
+            content: doc.content,
+          })),
         );
       });
     }
@@ -52,6 +66,18 @@ export default function ProfilePost() {
                     })
                   : 'No deadline'}
               </p>
+            </Link>
+          </li>
+        ))}
+      </ul>
+
+      <h1>내가 작성한 참여했Surv</h1>
+      <ul>
+        {userPostLite.map(post => (
+          <li key={post.id}>
+            <Link href={`/litesurveyposts/${post.id}`}>
+              {post.title}
+              {/* 필요한 경우 마감일 등 추가 정보 렌더링 */}
             </Link>
           </li>
         ))}
