+++ src/app/api/firebaseApi.ts
@@ -312,24 +312,3 @@ export const saveDataToFirebase = async (title: string, contents: string[], imag
 //     throw new Error('게시글을 불러오는 것에 실패했습니다.');
 //   }
 // };
-
-//프로필 페이지의 작성자가 작성한 IT 게시글
-export const getUserPostsIT = async (userId: string) => {
-  const postsQuery = query(collection(db, 'posts'), where('userId', '==', userId));
-  const querySnapshot = await getDocs(postsQuery);
-
-  const posts = querySnapshot.docs.map(doc => {
-    const docData = doc.data();
-    const deadlineDate = docData.deadlineDate ? docData.deadlineDate.toDate() : null;
-    console.log(doc);
-    console.log(doc.data);
-    return {
-      id: doc.id,
-      title: docData.title,
-      content: docData.content,
-      deadlineDate: deadlineDate,
-    };
-  });
-
-  return posts;
-};
